hubble/server: remove support for multiple listen addresses
Support for both a TCP and Unix Domain Socket based listeners does not
make sense as different options are likely required for different
listeners (eg: enable transport security with a TCP listener but not for
a Unix Domain Socket).
The possibility of configuring multiple listeners is actually not even
used by the Cilium daemon and 2 different Hubble server instances are
created to listen on both a Unix Domain Socket and a TCP socket.

Signed-off-by: Robin Hahling <robin.hahling@gw-computing.net>